third_party/WebKit/LayoutTests/inspector/console/console-copy-treeoutline-expected.txt

CONSOLE MESSAGE: line 11: [object Object]
Tests that console copies tree outline messages properly.


Running: testSelectAll
Selected text: console-copy-treeoutline.html:7 Object
